split lang and release controller tests
